♻️ refactor:   infrastructure/github/json_objects 配下のサブディレクトを廃止
Issue #157
