+++ lib/infrastructure/github/json_objects/owner.dart
