+++ lib/infrastructure/github/json_objects/owner.freezed.dart
