+++ lib/infrastructure/github/json_objects/search_repos_result.freezed.dart
