+++ test/infrastructure/github/json_objects/owner_test.dart
@@ -3,9 +3,9 @@
 // found in the LICENSE file.
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:github_search/infrastructure/github/json_objects/owner/owner.dart';
+import 'package:github_search/infrastructure/github/json_objects/owner.dart';
 
-import '../../../../test_utils/utils.dart';
+import '../../../test_utils/utils.dart';
 
 void main() {
   final ownerJsonObject = TestAssets.readJsonMap('github/owner.json')!;
